Move wolf integration tests to their own folder.
