+++ taskfiles/test.yaml
@@ -4,7 +4,7 @@ includes:
   toolchains: "toolchains.yaml"
 
 vars:
-  G_UNIT_TEST_BINARY: "{{.G_BUILD_SPIDER_DIR}}/tests/unitTest"
+  G_UNIT_TEST_BINARY: "{{.G_BUILD_SPIDER_DIR}}/tests/wolf/unitTest"
   G_TEST_VENV_DIR: "{{.G_BUILD_DIR}}/test-venv"
   G_TEST_VENV_CHECKSUM_FILE: "{{.G_BUILD_DIR}}/test#venv.md5"
 
@@ -202,7 +202,7 @@ tasks:
     cmds:
       - |-
         . {{.G_TEST_VENV_DIR}}/bin/activate
-        {{.G_TEST_VENV_DIR}}/bin/pytest tests/integration
+        {{.G_TEST_VENV_DIR}}/bin/pytest tests/wolf/integration
 
   # Internal task that runs all Spider Huntsman's unit tests.
   #
