--- tests/.clang-format
@@ -1,20 +1,0 @@
-BasedOnStyle: "InheritParentConfig"
-
-IncludeCategories:
-  # NOTE: A header is grouped by first matching regex library headers.
-  # NOTE: clang-format retains leading white-space on a line in violation of the YAML spec.
-  # Project library headers
-  - Regex: "^<(spider|tests)"
-    Priority: 4
-  # External library headers. Update when adding new libraries.
-  - Regex: "^<(antlr4|absl|boost|catch2|fmt|mariadb|msgpack|spdlog|ystdlib)"
-    Priority: 3
-  # C system headers
-  - Regex: "^<.+\\.h>"
-    Priority: 1
-  # C++ standard libraries
-  - Regex: "^<.+>"
-    Priority: 2
-  # Project relative headers
-  - Regex: "^\".+\""
-    Priority: 5
