+++ tests/wolf/scheduler/test-SchedulerServer.cpp
@@ -26,7 +26,7 @@
 #include <spider/storage/MetadataStorage.hpp>
 #include <spider/storage/StorageConnection.hpp>
 #include <spider/storage/StorageFactory.hpp>
-#include <tests/storage/StorageTestHelper.hpp>
+#include <tests/wolf/storage/StorageTestHelper.hpp>
 
 namespace {
 constexpr int cServerWarmupTime = 5;
