+++ tests/wolf/worker/signal-test.cpp
